// typings/contracts/PookyBallMinter.ts
export type MintTemplateStruct = {
  canMint: PromiseOrValue<boolean>;
  rarity: PromiseOrValue<BigNumberish>;
  maxMints: PromiseOrValue<BigNumberish>;
  currentMints: PromiseOrValue<BigNumberish>;
  price: PromiseOrValue<BigNumberish>;
  payingToken: PromiseOrValue<string>;
};